typelanguage: move parsing code to a module
