+++ src/type_language/mod.rs
@@ -1,1 +1,2 @@
 pub mod grammar;
+pub mod parser;
